pyproject.toml
[project]
name = "size-preserving-podcast-splicer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi[standard]>=0.115.7",
    "ffmpeg-python>=0.2.0",
    "pydub>=0.25.1",
]

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "pydub-stubs>=0.25.1.5",
]
